Add Tailwind CSS support and refactor recipe components
