+++ src/App.jsx
@@ -2,32 +2,61 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
+import { useRecettes } from './hooks/useRecettes';
+import AddRecette from './components/CrudRecette/AddRecette';
+import EdditRecette from './components/CrudRecette/EditRecette';
+import RemoveRecette from './components/CrudRecette/RemoveRecette';
 
 function App() {
   const [count, setCount] = useState(0)
+  const { recettes, addRecette, removeRecette, updateRecette, toggleFavorite } = useRecettes();
+  const [editingId, setEditingId] = useState(null);
+
+  function handleSaveEdit(updated) {
+    updateRecette(updated);
+    setEditingId(null);
+  }
 
   return (
     <>
-      <div>
-        <a href="https://vite.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.jsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
+
+      <div className="recette-app container p-4">
+          <h1>Mes recettes</h1>
+          <section className="mb-6">
+            <AddRecette addRecette={addRecette} />
+          </section>
+
+          <section>
+          <h1 class="text-3xl font-bold underline">
+          Hello world!
+          </h1>
+            {recettes.length === 0 && <div>Aucune recette pour l'instant.</div>}
+            <ul>
+              {recettes.map(r => (
+                <li key={r.id} className="recette-item mb-4 border p-3">
+                  {editingId === r.id ? (
+                    <EdditRecette recette={r} onSave={handleSaveEdit} onCancel={() => setEditingId(null)} />
+                  ) : (
+                    <div>
+                      <div className="flex justify-between items-start">
+                        <div>
+                          <h3>{r.title}</h3>
+                          <div className="text-sm text-gray-600">{(r.ingredients || []).join(', ')}</div>
+                        </div>
+                        <div className="space-x-2">
+                          <button onClick={() => toggleFavorite(r.id)}>{r.isFavorite ? '★' : '☆'}</button>
+                          <button onClick={() => setEditingId(r.id)}>Éditer</button>
+                          <RemoveRecette id={r.id} onRemove={() => removeRecette(r.id)} />
+                        </div>
+                      </div>
+                      <p className="mt-2">{r.preparation}</p>
+                    </div>
+                  )}
+                </li>
+              ))}
+            </ul>
+          </section>
+        </div>
     </>
   )
 }
