+++ src/components/CrudRecette/AddRecette.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { useRecettes } from '../hooks/useRecettes';
+import { useRecettes } from '../../hooks/useRecettes';
 
 export default function AddRecette({ addRecette: addRecetteProp }) {
   const hook = useRecettes();
